+++ src/main/java/com/sparta/mvm/service/PostService.java
@@ -9,6 +9,8 @@ import com.sparta.mvm.exception.ErrorEnum;
 import com.sparta.mvm.repository.PostRepository;
 import com.sparta.mvm.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,17 +23,10 @@ import java.util.List;
 public class PostService {
     private final PostRepository postRepository;
     private final UserRepository userRepository;
-
-    @Transactional(readOnly = true)
-    public PostResponseDto getPost(Long userId, PostRequestDto request) {
-        User user = getUserById(userId);
-        PostResponseDto responseDto = new PostResponseDto();
-        return responseDto;
-    }
-
+    
     public PostResponseDto findById(long postId) {
         Post post = findPostById(postId);
-        return PostResponseDto.toList("게시글 조회 성공 🎉", 200, post);
+        return PostResponseDto.toDto("게시글 조회 성공 🎉", 200, post);
     }
 
     private Post findPostById(long postId) {
@@ -39,8 +34,9 @@ public class PostService {
     }
 
     @Transactional
-    public PostResponseDto save(long userId, PostRequestDto request) {
-        User user = getUserById(userId);
+    public PostResponseDto save(PostRequestDto request) {
+        Long loggedInUserId = getLoggedInUserId();
+        User user = getUserById(loggedInUserId);
         Post post = request.toEntity();
         post.setUser(user);
         post = postRepository.save(post);
@@ -53,22 +49,32 @@ public class PostService {
         return list
                 .stream()
                 .sorted(Comparator.comparing(Post::getCreatedAt).reversed())
-                .map(post -> PostResponseDto.toList("게시글 조회 성공 🎉", 200, post))
+                .map(post -> PostResponseDto.toDto("게시글 조회 성공 🎉", 200, post))
                 .toList();
     }
 
     @Transactional
-    public PostResponseDto update( long postId, PostRequestDto request) {
+    public PostResponseDto update(long postId, PostRequestDto request) {
         Post post = findPostById(postId);
-        post.update(request.getContents());
-        return PostResponseDto.toDto("게시글 수정 성공 🎉", 200, post);
+        Long loggedInUserId = getLoggedInUserId();
+        if (loggedInUserId.equals(post.getUser().getId())) {
+            post.update(request.getContents());
+            return PostResponseDto.toDto("게시글 수정 성공 🎉", 200, post);
+        } else {
+            throw new CustomException(ErrorEnum.BAD_AUTH_PUT);
+        }
     }
 
     @Transactional
     public PostResponseDto delete(long postId) {
         Post post = findPostById(postId);
-        postRepository.delete(post);
-        return PostResponseDto.toDeleteResponse("게시글 삭제 성공 🎉", 200);
+        Long loggedInUserId = getLoggedInUserId();
+        if (loggedInUserId.equals(post.getUser().getId())) {
+            postRepository.delete(post);
+            return PostResponseDto.toDeleteResponse("게시글 삭제 성공 🎉", 200);
+        } else {
+            throw new CustomException(ErrorEnum.BAD_AUTH_DELETE);
+        }
     }
 
     private User getUserById(Long userId)
@@ -76,4 +82,12 @@ public class PostService {
         return userRepository.findById(userId).orElseThrow(() -> new CustomException(ErrorEnum.USER_NOT_FOUND));
     }
 
+    // 현재 로그인한 사용자의 ID 가져오기
+    private Long getLoggedInUserId() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+        User user = userRepository.findByUsername(username)
+                .orElseThrow(() -> new CustomException(ErrorEnum.USER_NOT_FOUND));
+        return user.getId();
+    }
 }
